✨ Added a flexible foundation for gadgets
A gadget is something that can create runtime runners that can then be ran to completion.

I've implemented a builder pipeline but I think it's a bit of a misuse of features to implement the constructed behaviours in feature callbacks. This seems like a service behaviour instead.

I might change this in the future.